% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congresses_and_years.R
\name{congress_in_year}
\alias{congress_in_year}
\title{Calculate the Congress number of a given year}
\usage{
congress_in_year(year)
}
\arguments{
\item{year}{Either a number or a Date object.
Cannot be earlier than 1789, the year of the first Congress.}
}
\value{
A positive whole number.
}
\description{
This function gives the number of the Congress for a specified calendar year.
}
\details{
A new Congress begins in every odd-numbered year, starting in 1789.
For example, 2021-2022 was the 117th Congress.
}
\examples{
congress_in_year(1800)
congress_in_year(2022)
}
