% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_voteview_cast_codes.R
\name{get_voteview_cast_codes}
\alias{get_voteview_cast_codes}
\title{Key to Voteview cast codes in individual member votes}
\usage{
get_voteview_cast_codes()
}
\value{
A tibble.
}
\description{
\code{get_voteview_cast_codes()} returns a tibble with definitions
of the 10 cast codes used in Voteview's member votes data
(i.e., the \code{cast_code} column in the data frames from
\code{\link[=get_voteview_member_votes]{get_voteview_member_votes()}}).
}
\details{
For more information on these cast codes, visit Voteview's
\href{https://voteview.com/articles/data_help_votes}{article} on the
member votes data.
}
\examples{
get_voteview_cast_codes()
}
\seealso{
\code{\link[=get_voteview_member_votes]{get_voteview_member_votes()}}, which uses these cast codes.
}
