% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filibustr-package.R
\docType{package}
\name{filibustr-package}
\alias{filibustr}
\alias{filibustr-package}
\title{filibustr: Data Utilities for Congressional Research}
\description{
The goal of 'filibustr' is to provide easy ways to access data on the U.S. Congress. This is originally designed for my B.A. honors thesis, but should be generally useful for congressional research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/feinleib/filibustr}
  \item Report bugs at \url{https://github.com/feinleib/filibustr/issues}
}

}
\author{
\strong{Maintainer}: Max Feinleib \email{mhfeinleib@gmail.com} (\href{https://orcid.org/0009-0002-9604-3533}{ORCID})

}
\keyword{internal}
