% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filibustr-package.R
\docType{package}
\name{filibustr-package}
\alias{filibustr}
\alias{filibustr-package}
\title{filibustr: Data Utilities for Congressional Research}
\description{
Provides easy-to-understand and consistent interfaces for accessing data on the U.S. Congress.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/feinleib/filibustr}
  \item Report bugs at \url{https://github.com/feinleib/filibustr/issues}
}

}
\author{
\strong{Maintainer}: Max Feinleib \email{mhfeinleib@gmail.com} (\href{https://orcid.org/0009-0002-9604-3533}{ORCID}) [copyright holder]

}
\keyword{internal}
