% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senate_cloture_votes.R
\name{get_senate_cloture_votes}
\alias{get_senate_cloture_votes}
\title{Senate cloture motions and votes}
\usage{
get_senate_cloture_votes()
}
\value{
A \code{\link[=tibble]{tibble()}} with the number of cloture motions, cloture votes, and
successful cloture votes in each Congress.
}
\description{
\code{get_senate_cloture_votes()} returns a tibble of the number of cloture motions,
cloture votes, and successful cloture votes in the Senate during each Congress
since 1917.
}
\details{
The data is sourced from the official Senate website, specifically
\url{https://www.senate.gov/legislative/cloture/clotureCounts.htm}.
}
\examples{
get_senate_cloture_votes()
}
