% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_voteview_members.R
\name{get_voteview_members}
\alias{get_voteview_members}
\title{Get data on members of Congress from Voteview}
\usage{
get_voteview_members(local = TRUE, local_dir = ".", chamber = "all")
}
\arguments{
\item{local}{Whether to read the data from a local file, as opposed to the Voteview website.
Default is \code{TRUE}. If the local file does not exist, will fall back to reading from Voteview.}

\item{local_dir}{The directory containing the local file. Defaults to the working directory.}

\item{chamber}{Which chamber to get data for. Options are:
\itemize{
\item \code{"all"}, \code{"congress"}: Both House and Senate data (the default).
\item \code{"house"}, \code{"h"}, \code{"hr"}: House data only.
\item \code{"senate"}, \code{"s"}, \code{"sen"}: Senate data only.
}

Note that presidents are included in all datasets. Therefore, reading \emph{both} \code{"house"}
and \code{"senate"} data will duplicate data on the presidents. The recommended way to get
all data is to use the default argument, \code{"all"}.}
}
\value{
A \code{\link[=tibble]{tibble()}}.

The tibble includes data on the member's office, party, and ideology.
See \href{https://voteview.com/data}{Voteview} for descriptions of specific columns.
}
\description{
\code{get_voteview_members()} returns a tibble of data on members of Congress,
sourced from \href{https://voteview.com/data}{Voteview}. Members in the data include
Senators, Representatives, and Presidents. Each row is one member in one
Congress (i.e., each member is listed once for every two years in office). See the
\href{https://voteview.com/data}{Voteview} website for more information on their data.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_voteview_members()

# Get data for only one chamber
get_voteview_members(chamber = "house")
get_voteview_members(chamber = "senate")
\dontshow{\}) # examplesIf}
# Get data from Voteview website
try(get_voteview_members(local = FALSE))
}
