% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_voteview_parties.R
\name{get_voteview_parties}
\alias{get_voteview_parties}
\title{Get ICPSR party codes from Voteview}
\usage{
get_voteview_parties(
  local = TRUE,
  local_dir = ".",
  chamber = "all",
  congress = NULL
)
}
\arguments{
\item{local}{Whether to read the data from a local file, as opposed to the Voteview website.
Default is \code{TRUE}. If the local file does not exist, will fall back to reading from Voteview.}

\item{local_dir}{The directory containing the local file. Defaults to the working directory.}

\item{chamber}{Which chamber to get data for. Options are:
\itemize{
\item \code{"all"}, \code{"congress"}: Both House and Senate data (the default).
\item \code{"house"}, \code{"h"}, \code{"hr"}: House data only.
\item \code{"senate"}, \code{"s"}, \code{"sen"}: Senate data only.
These options are case-insensitive. If you explicitly pass a different value,
it will default to "all" with a warning.
}

Note that presidents are included in all datasets. Therefore, reading \emph{both} \code{"house"}
and \code{"senate"} data will duplicate data on the presidents. The recommended way to get
all data is to use the default argument, \code{"all"}.}

\item{congress}{A single whole number (to get data for a single Congress), or
a numeric vector (to get data for a set of congresses).
Optional; will retrieve data for all Congresses by default.
If specified, Congress numbers cannot be greater than the \code{\link[=current_congress]{current_congress()}}
(i.e., you cannot try to get future data).}
}
\value{
A \code{\link[=tibble]{tibble()}}.
}
\description{
Get ICPSR party codes from Voteview
}
\examples{
get_voteview_parties()
}
